Move reverb code out of utils
