+++ Jamulus.pro
@@ -378,7 +378,8 @@ FORMS_GUI = src/aboutdlgbase.ui \
         src/connectdlgbase.ui
 }
 
-HEADERS += src/buffer.h \
+HEADERS += src/plugins/audioreverb.h \
+    src/buffer.h \
     src/channel.h \
     src/global.h \
     src/protocol.h \
@@ -486,7 +487,8 @@ HEADERS_OPUS_X86 = libs/opus/celt/x86/celt_lpc_sse.h \
     libs/opus/celt/x86/x86cpu.h \
     $$files(libs/opus/silk/x86/*.h)
 
-SOURCES += src/buffer.cpp \
+SOURCES += src/plugins/audioreverb.cpp \
+    src/buffer.cpp \
     src/channel.cpp \
     src/main.cpp \
     src/protocol.cpp \
